Add routings constraints files
