+++ config/routes.rb
@@ -1,22 +1,4 @@
-class PageCountConstraint
-  PAGE_COUNT_LIMIT = 100_000
-
-  def self.matches?(request)
-    return true unless request.query_parameters['page']
-    request.query_parameters['page'].to_i < PAGE_COUNT_LIMIT
-  end
-end
-
-class TargetYearConstraint
-  YEARS_LOWER_LIMIT = 2000
-  YEARS_UPPER_LIMIT = 2100
-
-  def self.matches?(request)
-    target_year = request.query_parameters['target_year']&.to_i
-    return true unless target_year
-    (target_year >= YEARS_LOWER_LIMIT) && (target_year < YEARS_UPPER_LIMIT)
-  end
-end
+require_relative 'routes/constraints'
 
 Rails.application.routes.draw do
 
@@ -36,12 +18,12 @@ Rails.application.routes.draw do
   resources :groups, only: [:index]
   resources :monthly_report_comments, only: [:create, :edit, :update, :destroy]
 
-  root to: 'monthly_reports#index', constraints: PageCountConstraint
+  root to: 'monthly_reports#index', constraints: Constraints::PageCount
 
   resources :user_profiles, only: [:show, :edit, :update]
-  resources 'monthly_reports', except: :destroy, constraints: PageCountConstraint do
+  resources 'monthly_reports', except: :destroy, constraints: Constraints::PageCount do
     collection do
-      get 'users/:user_id', action: :user, as: :user, user_id: /\d{,6}/, constraints: TargetYearConstraint
+      get 'users/:user_id', action: :user, as: :user, user_id: /\d{,6}/, constraints: Constraints::TargetYear
       get :copy
     end
   end
